+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Change the formatting of Backtraces and how they are customized [#44](https://github.com/rootcause-rs/rootcause/pull/44)
+- Change the formatting of Backtraces and Location and how they are customized [#44](https://github.com/rootcause-rs/rootcause/pull/44)
 
 ## [0.7.0] - 2025-11-06