+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0] - Unreleased
+
+### Changed
+
+- Change the formatting of Backtraces and Location and how they are customized [#44](https://github.com/rootcause-rs/rootcause/pull/44)
 
 ## [0.7.0] - 2025-11-06
 
@@ -89,7 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/rootcause-rs/rootcause/compare/v0.7.0...HEAD
+[0.8.0]: https://github.com/rootcause-rs/rootcause/compare/v0.7.0...HEAD
 [0.7.0]: https://github.com/rootcause-rs/rootcause/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/rootcause-rs/rootcause/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/rootcause-rs/rootcause/compare/v0.4.3...v0.5.0